api_access_control/requirements.txt
asgiref==3.8.1
autopep8==2.3.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cloudinary==1.41.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.1.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.5.0
django-filter==24.3
django-import-export==3.0.2
django-livereload-server==0.5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
et-xmlfile==1.1.0
gunicorn==23.0.0
holidays==0.65
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
Markdown==3.7
MarkupPy==1.14
MarkupSafe==3.0.2
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.1
pillow==11.0.0
psycopg2==2.9.10
pycodestyle==2.12.1
pycparser==2.22
PyJWT==2.9.0
pypng==0.20220715.0
PyQRCode==1.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
setuptools==75.8.0
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
tornado==6.4.1
uritemplate==4.1.1
urllib3==2.2.3
wheel==0.45.1
whitenoise==6.8.2
xlrd==2.0.1
xlwt==1.3.0
